Precalculated Data Storage

Raw Data

Tables of data can be found here:

Table of primes Hecke operators on modular forms modulo two.

Table of powers of Hecke operators on modular forms modulo two.

If this is not enought, remeber that this module allows anyone to generate new data.

Use of Precalculated Variables

Here is how to use the variables:

$\ T_p|\Delta^k$ is Hecke_primes[p][k+1]

$\ T_3^iT_5^j|\Delta^k$ is Hecke_powers[i+1,j+1][k+1]

These are stored as Δ-series. The q-series of powers of Δ are stored as follows:

$\ \Delta^k$ is precalculated[k+1]

Generating Precalculated Data

The various precalculated data generators may be found in the data subfolder. Note that the user may create new data files, if the provided ones aren't enought. Note as well that there are two implemented ways to store data, we advice the binary for speed purposes.

Binary Data Reads

Main.ModularFormsModuloTwo.loadFormListBinary — Method

loadFormListBinary(MAXI, LENGTH)

Loads the list of q-coefficients of Δ powers form file.

Main.ModularFormsModuloTwo.loadHeckePowersListBinary — Method

loadHeckePowersListBinary(MAX_POWER, MAX_DELTA)

Loads the list of Δ-coefficients of powers of Hecke operators applied to powers of Δ.

Main.ModularFormsModuloTwo.loadHeckePrimesListBinary — Method

loadHeckePrimesListBinary(MAX_PRIME, MAX_DELTA)

Loads the list of Δ-coefficients of prime Hecke operators applied to powers of Δ.

Text Data Reads

Main.ModularFormsModuloTwo.loadForm — Function

loadForm(name, file_name)

Load the modular form modulo 2 name from file_name.

Main.ModularFormsModuloTwo.loadFormList — Function

loadFormList(file_name)

Load the modular form list from file_name.

Main.ModularFormsModuloTwo.saveForm — Function

saveForm(f, name, file_name)

Save the modular form modulo 2 f as name in file_name.

Main.ModularFormsModuloTwo.saveFormList — Function

saveFormList(forms_list, file_name)

Save the modular forms modulo 2 f as name in file_name.